se implementa el metodo MVVM para recuperar la informacion de lossensores almacenada en Firebase

+++ app/src/main/java/com/daferarevalo/espapp/ui/channels/ChannelsFragment.kt
@@ -2,11 +2,9 @@ package com.daferarevalo.espapp.ui.channels
 
 import android.os.Bundle
 import android.util.Log
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.widget.Toast
+import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.Observer
 import androidx.navigation.fragment.findNavController
@@ -47,7 +45,7 @@ class ChannelsFragment : Fragment(R.layout.fragment_channels) {
                 is Result.Success -> {
                     binding.progressBar.hide()
                     binding.addChanelButton.isEnabled = true
-                    ch = result.data
+                    ch = result.data.toInt()
                     //Toast.makeText(context,"Bienvenido $")
                     when(ch){
                         1-> {
